refactor(acp): split native result projection
