+++ src/acp/native-agent.ts
@@ -24,9 +24,7 @@ import {
 import { agentCommandFromIngress } from "../agents/agent-command.js";
 import { resolveDefaultAgentId } from "../agents/agent-scope-config.js";
 import { resolveEmbeddedAbortSettleTimeoutMs } from "../agents/embedded-agent-runner/run/attempt.abort-settle-timeout.js";
-import { stripInternalRuntimeContext } from "../agents/internal-runtime-context.js";
 import type { LocalExecApprovalRequest } from "../agents/local-exec-approval-broker.js";
-import { isSilentReplyPayloadText } from "../auto-reply/tokens.js";
 import { getRuntimeConfig } from "../config/config.js";
 import { type AgentEventPayload, onAgentEvent } from "../infra/agent-events.js";
 import { buildApprovalPresentation } from "../infra/approval-presentation.js";
@@ -50,9 +48,9 @@ import type {
 } from "../infra/plugin-approvals.js";
 import { toAgentStoreSessionKey } from "../routing/session-key.js";
 import type { RuntimeEnv } from "../runtime.js";
-import { stripInlineDirectiveTagsForDisplay } from "../utils/directive-tags.js";
 import { INTERNAL_MESSAGE_CHANNEL } from "../utils/message-channel.js";
 import { extractAttachmentsFromPrompt, extractTextFromPrompt } from "./event-mapper.js";
+import { agentResultError, finalVisibleText, type AgentResult } from "./native-agent-result.js";
 import {
   buildAcpPermissionRequest,
   resolveGatewayDecisionFromPermissionOutcome,
@@ -72,7 +70,6 @@ const silentRuntime: RuntimeEnv = {
 };
 
 type AgentExecutor = typeof agentCommandFromIngress;
-type AgentResult = Awaited<ReturnType<AgentExecutor>>;
 
 type NativeSession = {
   id: string;
@@ -99,66 +96,6 @@ type NativeAgentDependencies = {
   };
   abortSettleTimeoutMs?: number;
 };
-
-function payloadText(parts: unknown): string {
-  if (!Array.isArray(parts)) {
-    return "";
-  }
-  return parts
-    .flatMap((part) => {
-      if (!part || typeof part !== "object") {
-        return [];
-      }
-      const payload = part as {
-        isCommentary?: unknown;
-        isError?: unknown;
-        isReasoning?: unknown;
-        text?: unknown;
-        visible?: unknown;
-      };
-      return payload.isCommentary !== true &&
-        payload.isError !== true &&
-        payload.isReasoning !== true &&
-        payload.visible !== false &&
-        typeof payload.text === "string"
-        ? [payload.text]
-        : [];
-    })
-    .filter(Boolean)
-    .join("\n\n");
-}
-
-function finalVisibleText(result: AgentResult | undefined): string {
-  const text = result?.meta.finalAssistantVisibleText ?? payloadText(result?.payloads);
-  const normalized = stripInternalRuntimeContext(stripInlineDirectiveTagsForDisplay(text).text);
-  return isSilentReplyPayloadText(normalized) ? "" : normalized;
-}
-
-function agentResultError(result: AgentResult | undefined): Error | undefined {
-  if (!result) {
-    return new Error("OpenClaw agent returned no result");
-  }
-  const errorPayload = result.payloads?.find(
-    (payload) => (payload as { isError?: unknown }).isError === true,
-  ) as { text?: unknown } | undefined;
-  const message =
-    result.meta.error?.message ??
-    (typeof errorPayload?.text === "string" && errorPayload.text.trim()
-      ? errorPayload.text.trim()
-      : undefined);
-  if (result.meta.stopReason === "timeout" || result.meta.timeoutPhase !== undefined) {
-    return new Error(message ?? "OpenClaw agent timed out");
-  }
-  if (
-    result.meta.aborted === true ||
-    result.meta.stopReason === "error" ||
-    result.meta.error !== undefined ||
-    errorPayload
-  ) {
-    return new Error(message ?? "OpenClaw agent run failed");
-  }
-  return undefined;
-}
 
 function permissionOptions(decisions: readonly ExecApprovalDecision[]): PermissionOption[] {
   return decisions.map((decision) => ({
